+++ QueenGlandBuff/MainPlugin.cs
@@ -21,7 +21,7 @@ namespace QueenGlandBuff
 	{
 		public const string MODUID = "com.kking117.QueenGlandBuff";
 		public const string MODNAME = "QueenGlandBuff";
-		public const string MODVERSION = "1.5.3";
+		public const string MODVERSION = "1.5.4";
 
 		public const string MODTOKEN = "KKING117_QUEENGLANDBUFF_";
 
@@ -53,6 +53,10 @@ namespace QueenGlandBuff
 		private void PostLoad()
 		{
 			//LoadCompat();
+			if (Changes.QueensGland.Enable)
+			{
+				Changes.QueensGland.PostLoad();
+			}
 			BazaarSceneDef = SceneCatalog.FindSceneDef("bazaar");
 		}
 	}